% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libproj-package.R
\docType{package}
\name{libproj-package}
\alias{libproj}
\alias{libproj-package}
\title{libproj: Generic Coordinate Transformation Library ('PROJ') C API}
\description{
Provides a 'PROJ' <https://proj.org> C API that can be 
  used to write high-performance C and C++ coordinate transformation
  operations using R as an interface. This package contains an internal 
  version of the 'PROJ' library to guarantee the best possible 
  consistency on multiple platforms, and to provide a means by which
  'PROJ' can be used on platforms where it may be impractical
  or impossible to install a binary version of the library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/libproj}
  \item Report bugs at \url{https://github.com/paleolimbot/libproj/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item PROJ authors
}

Other contributors:
\itemize{
  \item Jeroen Ooms (configure script) [copyright holder]
  \item Evan Roualt [copyright holder]
  \item Thomas Knudsen [copyright holder]
  \item SDFE [copyright holder]
  \item Kristan Evers [copyright holder]
  \item Frank Warmerdam [copyright holder]
  \item Charles Karney [copyright holder]
  \item Evan Roualt [copyright holder]
}

}
\keyword{internal}
