% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libproj-package.R
\name{libproj_cleanup}
\alias{libproj_cleanup}
\title{Clean up resources used by the default context}
\usage{
libproj_cleanup()
}
\value{
NULL, invisibly
}
\description{
Clean up resources used by the default context
}
\examples{
libproj_cleanup()

}
