% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libprojmds-trans_xy.R
\name{proj_create}
\alias{proj_create}
\title{Create CRS string}
\usage{
proj_create(x, format = 0L)
}
\arguments{
\item{x}{input crs string}

\item{format}{(0 is WKT2)}
}
\value{
character string
}
\description{
Create CRS string
}
\examples{
proj_create("+proj=longlat +datum=WGS84")
}
