% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libprojmds-trans_xy.R
\name{proj_trans_xy}
\alias{proj_trans_xy}
\title{Generic x/y coordinate transformation, source to target}
\usage{
proj_trans_xy(x, y, source, target)
}
\arguments{
\item{x}{x coordinates (longitude or x)}

\item{y}{y coordinates (latitude or y)}

\item{source}{crs input projection string (EPSG:CODE not working rn, wkt or proj, urn ok)}

\item{target}{crs output projection string}
}
\value{
list of transformed x, y
}
\description{
Generic x/y coordinate transformation, source to target
}
\examples{
proj_trans_xy(147, -42, source = "+proj=longlat +datum=WGS84", target = "+proj=merc")
}
