% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libproj-package.R
\name{libproj_version}
\alias{libproj_version}
\alias{libproj_has_libtiff}
\alias{libproj_has_libcurl}
\alias{libproj_temp_dir}
\alias{libproj_configuration}
\alias{libproj_configure}
\title{PROJ configuration}
\usage{
libproj_version()

libproj_has_libtiff()

libproj_has_libcurl()

libproj_temp_dir()

libproj_configuration()

libproj_configure(
  search_path = c(system.file("proj", package = "libproj"),
    getOption("libproj.search_path", NULL)),
  db_path = getOption("libproj.db_path", character(0)),
  user_writable_dir = getOption("libproj.user_writable_dir", libproj_temp_dir()),
  network_endpoint = getOption("libproj.network_endpoint", "https://cdn.proj.org")
)
}
\arguments{
\item{search_path}{A character vector of paths where libproj will
look for resource files (e.g., gridshift, database, init files).
Defaults to the internal database and init files distributed
with the PROJ source.}

\item{db_path}{Independent of \code{search_path}, a character vector of
SQLite databases that may contain updated or alternative
coordinate operations from the default proj.db included in this
package. By default, this defers to the search path.}

\item{user_writable_dir}{A directory that can safely be written to
by this package. This contains a cache of grid shift files downloaded
from the PROJ CDN at \code{network_endpoint} if using \code{with_libproj_network()}.}

\item{network_endpoint}{A mirror of the PROJ CDN of gridshift files. By default,
this is set to \url{https://cdn.proj.org}.}
}
\description{
PROJ configuration
}
\examples{
libproj_version()
libproj_has_libtiff()
libproj_has_libcurl()
libproj_temp_dir()
libproj_configuration()


}
